An unexpected behaviour of wiringPi interrupts running in another thread fixed.
Signals under linux are raised to just one thread, which meant that the main thread, under which sigsuspend is waiting, would not be alerted to SIGUSR1 being raised and thus never show the video.  
On another front, attempting to change the command to take a picture so that it doesn't have to be queued after the video has finished, by instead linking the last frame. Needs to be tested.